Refactored format text
